fl2-909 SNPs outside coordinate range returned

+++ src/main/java/org/jax/mgi/snpdatamodel/ConsensusCoordinateSNP.java
@@ -12,7 +12,7 @@ import lombok.EqualsAndHashCode;
 public class ConsensusCoordinateSNP extends AbstractSNP {
 	
 	private String chromosome;
-	private int startCoordinate;
+	private long startCoordinate;
 	private boolean multiCoord;
 	private String strand;
