+++ src/main/java/org/jax/mgi/snpdatamodel/document/SearchSNPDocument.java
@@ -10,7 +10,7 @@ import lombok.EqualsAndHashCode;
 public class SearchSNPDocument extends BaseESDocument {
 
 	private String chromosome;
-	private Double startcoordinate;
+	private Long startcoordinate;
 	private String varclass;
 	private List<String> fxn;
 	private List<String> marker_accid;
